Consolidate Geant4 geometry conversion utilities
